+++ HISTORY.md
@@ -1,6 +1,6 @@
 0.2.3 / 2-12-2013
 =============
-* Adding check for count in properties length.
+* Adding check for count in properties and traits length.
 
 0.2.2 / 2-11-2013
 =============